+++ layers/CMSSchema/packages/users/src/FieldResolvers/ObjectType/AbstractUserObjectTypeFieldResolver.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace PoPCMSSchema\Users\FieldResolvers\ObjectType;
 
 use PoPCMSSchema\SchemaCommons\DataLoading\ReturnTypes;
-use PoPCMSSchema\SchemaCommons\Resolvers\WithLimitFieldArgResolverTrait;
 use PoPCMSSchema\Users\TypeAPIs\UserTypeAPIInterface;
 use PoPCMSSchema\Users\TypeResolvers\InputObjectType\UserPaginationInputObjectTypeResolver;
 use PoPCMSSchema\Users\TypeResolvers\InputObjectType\UserSortInputObjectTypeResolver;
@@ -24,8 +23,6 @@ use PoP\GraphQLParser\Spec\Parser\Ast\FieldInterface;
 
 abstract class AbstractUserObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolver
 {
-    use WithLimitFieldArgResolverTrait;
-
     private ?UserTypeAPIInterface $userTypeAPI = null;
     private ?IntScalarTypeResolver $intScalarTypeResolver = null;
     private ?UserObjectTypeResolver $userObjectTypeResolver = null;