+++ layers/GatoGraphQLForWP/plugins/gatographql/src/ModuleResolvers/Extensions/PremiumBundleExtensionModuleResolver.php
@@ -16,6 +16,7 @@ class PremiumBundleExtensionModuleResolver extends AbstractBundleExtensionModule
     public const BRICKS = Plugin::NAMESPACE . '\\bundle-extensions\\bricks';
     public const ELEMENTOR = Plugin::NAMESPACE . '\\bundle-extensions\\elementor';
     public const EVENTS_MANAGER = Plugin::NAMESPACE . '\\bundle-extensions\\events-manager';
+    public const JETENGINE = Plugin::NAMESPACE . '\\bundle-extensions\\jetengine';
     public const MULTILINGUALPRESS = Plugin::NAMESPACE . '\\bundle-extensions\\multilingualpress';
     public const POLYLANG = Plugin::NAMESPACE . '\\bundle-extensions\\polylang';
     public const TRANSLATION = Plugin::NAMESPACE . '\\bundle-extensions\\translation';
@@ -34,6 +35,7 @@ class PremiumBundleExtensionModuleResolver extends AbstractBundleExtensionModule
             self::BRICKS,
             self::ELEMENTOR,
             self::EVENTS_MANAGER,
+            self::JETENGINE,
             self::MULTILINGUALPRESS,
             self::POLYLANG,
             self::TRANSLATION,
@@ -48,6 +50,7 @@ class PremiumBundleExtensionModuleResolver extends AbstractBundleExtensionModule
             self::BRICKS => \__('Bricks', 'gatographql'),
             self::ELEMENTOR => \__('Elementor', 'gatographql'),
             self::EVENTS_MANAGER => \__('Events Manager', 'gatographql'),
+            self::JETENGINE => \__('JetEngine', 'gatographql'),
             self::MULTILINGUALPRESS => \__('MultilingualPress', 'gatographql'),
             self::POLYLANG => \__('Polylang', 'gatographql'),
             self::TRANSLATION => \__('Translation', 'gatographql'),
@@ -63,6 +66,7 @@ class PremiumBundleExtensionModuleResolver extends AbstractBundleExtensionModule
             self::BRICKS => \__('Integration with plugin "Bricks", adding fields to parse and update data in Bricks pages and templates.', 'gatographql'),
             self::ELEMENTOR => \__('Integration with plugin "Elementor", adding fields to parse and update data in Elementor pages and templates.', 'gatographql'),
             self::EVENTS_MANAGER => \__('Integration with plugin "Events Manager", adding fields to the schema to fetch event data.', 'gatographql'),
+            self::JETENGINE => \__('Integration with plugin "JetEngine", adding fields to the schema to fetch Custom Content Type data.', 'gatographql'),
             self::MULTILINGUALPRESS => \__('Integration with plugin "MultilingualPress", adding fields to the schema to fetch multilingual data.', 'gatographql'),
             self::POLYLANG => \__('Integration with plugin "Polylang", adding fields to the schema to fetch multilingual data.', 'gatographql'),
             self::TRANSLATION => \__('Translate content to multiple languages using the service provider of your choice, among ChatGPT, Claude, DeepSeek, Mistral AI, DeepL, and Google Translate.', 'gatographql'),
@@ -85,6 +89,7 @@ class PremiumBundleExtensionModuleResolver extends AbstractBundleExtensionModule
             self::BRICKS => $imagePathURL . '/bricks.svg',
             self::ELEMENTOR => $imagePathURL . '/elementor.svg',
             self::EVENTS_MANAGER => $imagePathURL . '/events-manager.webp',
+            self::JETENGINE => $imagePathURL . '/jetengine.svg',
             self::MULTILINGUALPRESS => $imagePathURL . '/multilingualpress.webp',
             self::POLYLANG => $imagePathURL . '/polylang.webp',
             self::TRANSLATION => $imagePathURL . '/translation.svg',
@@ -110,6 +115,9 @@ class PremiumBundleExtensionModuleResolver extends AbstractBundleExtensionModule
             ],
             self::EVENTS_MANAGER => [
                 PremiumExtensionModuleResolver::EVENTS_MANAGER,
+            ],
+            self::JETENGINE => [
+                PremiumExtensionModuleResolver::JETENGINE_CCTS,
             ],
             self::MULTILINGUALPRESS => [
                 PremiumExtensionModuleResolver::MULTILINGUALPRESS,